Tooltipのファイルを整理
